refactor: calculateGap 반환 타입 변경

+++ matzip-app-external-api/src/main/java/com/woowacourse/matzip/application/ReviewService.java
@@ -63,7 +63,7 @@ public class ReviewService {
 
         Review review = reviewRepository.findById(reviewId)
                 .orElseThrow(ReviewNotFoundException::new);
-        long reviewGap = review.calculateGap(reviewUpdateRequest.getRating());
+        int reviewGap = review.calculateGap(reviewUpdateRequest.getRating());
         review.update(member.getGithubId(),
                 reviewUpdateRequest.getContent(),
                 reviewUpdateRequest.getRating(),
